claim_winnings + update state
- claim_winnings now supports tokens
- update_state module added with ix handlers to update various state
fields
- removed update_withdraw_unclaimed module and moved logic into
update_state

+++ programs/santa-vs-grinch/src/state/config.rs
@@ -22,7 +22,7 @@ pub struct InitializeArgs {
     pub admin_fee_percentage_bp: u16,
     pub bet_burn_percentage_bp: u16,
     pub mystery_box_burn_percentage_bp: u16,
-    pub mystery_box_price: u32,
+    pub mystery_box_price: u64,
     pub creators: [Creator; MAX_CREATORS],
 }
 
@@ -36,7 +36,7 @@ pub struct Config {
     pub bet_burn_percentage_bp: u16,
     pub mystery_box_burn_percentage_bp: u16,
 
-    pub mystery_box_price: u32,
+    pub mystery_box_price: u64,
 
     pub vault: Pubkey,
     pub fees_vault: Pubkey,
